Rename `parse` to `parser`
