+++ src/ide/doc.ts
@@ -4,8 +4,8 @@ import { ErrorHelper } from "../lang/errors/ErrorHelper";
 import { tgui } from "./tgui";
 import { defaultOptions } from "../lang/helpers/options";
 import { Interpreter } from "../lang/interpreter/interpreter";
-import { Lexer } from "../lang/parse/lexer";
-import { Parser } from "../lang/parse/parser";
+import { Lexer } from "../lang/parser/lexer";
+import { Parser } from "../lang/parser";
 import { TScript } from "../lang";
 import { Version } from "../lang/version";
 import { searchengine } from "./search";
