+++ src/interop.ts
@@ -3,8 +3,8 @@ import { defaultOptions } from "./lang/helpers/options";
 import { Typeid } from "./lang/helpers/typeIds";
 import { defaultService } from "./lang/interpreter/defaultService";
 import { Interpreter as _Interpreter } from "./lang/interpreter/interpreter"
-import { Lexer } from "./lang/parse/lexer";
-import { Parser } from "./lang/parse/parser"
+import { Lexer } from "./lang/parser/lexer";
+import { Parser } from "./lang/parser"
 import { TScript } from "./lang"
 import { Version } from "./lang/version";
