+++ src/lang/interpreter/interpreter_helper.ts
@@ -1,5 +1,5 @@
 import { ErrorHelper } from "../errors/ErrorHelper";
-import { binary_operator_impl, left_unary_operator_impl } from "../parse/parser_helper";
+import { binary_operator_impl, left_unary_operator_impl } from "../parser/parser_helper";
 import { TScript } from "..";
 import { simfalse } from "../helpers/sims";
 import { Typeid } from "../helpers/typeIds";
