+++ src/lang/parser/parse_declaration.ts
