+++ src/lang/parser/parse_namespace.ts
